pyproject.toml
[tool.poetry]
name = "video-gem"
version = "0.1.0"
description = "A video processing application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "*"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
requests = "^2.28.1"
python-multipart = "^0.0.5"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
openai = "^1.82.0"
elevenlabs = "^2.1.0"
opencv-python-headless = "^4.11.0.86"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
